# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Helix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings/": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "description": "Return current configuration with masked API keys.\n\nReads DB Setting rows (including encrypted API keys) and merges onto\nGeneConfig defaults. Env var keys take priority over DB-stored keys.",
        "operationId": "get_settings_api_settings__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Settings",
        "description": "Update configuration in database and return new settings.\n\nOnly provided (non-None) fields are updated; others are preserved.\nAPI key fields are encrypted and stored in the \"api_keys\" Setting category.\nNon-key fields go to \"global_config\" as before.\nThe config cache is cleared so the next request picks up new values.",
        "operationId": "update_settings_api_settings__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/test-connection": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Test Connection",
        "description": "Validate an API key by making a lightweight request to the provider.\n\nAttempts to list models from the provider. Returns success/failure.",
        "operationId": "test_connection_api_settings_test_connection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleConfig": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "has_key": {
            "type": "boolean",
            "title": "Has Key"
          },
          "key_hint": {
            "type": "string",
            "title": "Key Hint",
            "default": ""
          },
          "thinking_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Budget"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "has_key"
        ],
        "title": "RoleConfig",
        "description": "Configuration for a single role (meta/target/judge)."
      },
...
      "SettingsResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/RoleConfig"
          },
          "target": {
            "$ref": "#/components/schemas/RoleConfig"
          },
          "judge": {
            "$ref": "#/components/schemas/RoleConfig"
          },
          "concurrency_limit": {
            "type": "integer",
            "title": "Concurrency Limit"
          },
          "generation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Providers"
          },
          "has_db_keys": {
            "type": "boolean",
            "title": "Has Db Keys",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "meta",
          "target",
          "judge",
          "concurrency_limit",
          "generation",
          "providers"
        ],
        "title": "SettingsResponse",
        "description": "Response from GET /api/settings."
      },
      "SettingsUpdateRequest": {
        "properties": {
          "meta_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Provider"
          },
          "meta_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Model"
          },
          "target_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Provider"
          },
          "target_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Model"
          },
          "judge_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Provider"
          },
          "judge_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Model"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          },
          "gemini_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Api Key"
          },
          "openai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Key"
          },
          "concurrency_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency Limit"
          },
          "generation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation"
          },
          "meta_thinking_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Thinking Budget"
          },
          "target_thinking_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Thinking Budget"
          },
          "judge_thinking_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Thinking Budget"
          }
        },
        "type": "object",
        "title": "SettingsUpdateRequest",
        "description": "Request body for PUT /api/settings. All fields optional for partial update."
      },
...
      "TestConnectionRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "api_key"
        ],
        "title": "TestConnectionRequest",
        "description": "Request body for POST /api/settings/test-connection."
      },
      "TestConnectionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "TestConnectionResponse",
        "description": "Response from POST /api/settings/test-connection."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}